// src/app/features/models/status.model.ts
export interface StatusResponse {
  success: boolean;
  data: [];
  message: string;
}

export interface StatusList {
  _id: string;
  name: string;
  code: string;
  color: string;
  description: string;
  isCustomerVisible: boolean;
  isSellable: boolean;
  sortOrder: number;
  isActive: boolean;
  createdBy: string;
  createdOn: string;
  updatedOn: string;
}

export interface Status {
  name: string;
  code: string;
  isActive: boolean;
}
export interface StatusInfo {
   name: string;
  code: string;
  color: string;
  description: string;
  isCustomerVisible: boolean;
  isSellable: boolean;
  sortOrder: number;
  isActive: boolean;
  createdBy: {
    _id: string;
    firstName: string;
    lastName: string;
    email: string;
  };
  createdOn: string;
  updatedOn: string;
}